erp_core/requirements.txt
fastapi==0.115.8
numpy==2.0.2
pandas==2.2.3
pytz==2024.2
scikit-learn==1.6.1
scipy==1.13.1
uvicorn==0.34.0
django==4.2.18
djangorestframework~=3.15.2
psycopg2==2.9.10
drf-yasg~=1.21.8
django-environ~=0.12.0
django-filter~=24.3